Login y Registro

+++ src/App/Controlador/ControladorPagina.php
@@ -132,6 +132,12 @@ class ControladorPagina{
         require $this->viewsDir . 'mi-cuenta.view.php';
     }
 
+    public function registro(){
+        $titulo = 'PawPrints - Registro';
+        $htmlClass = "mi-cuenta-pages";
+        require $this->viewsDir . 'registro.view.php';
+    }
+
     public function quienesSomos(){
         $titulo = 'PawPrints - Quiénes somos';
         $htmlClass = "preguntas-pages";
@@ -196,6 +202,65 @@ class ControladorPagina{
         
         // Ruta del archivo de texto donde se guardarán los datos
         $archivo = __DIR__ . "/../../reservas.txt";
+        
+        // Guardar los datos en el archivo de texto
+        file_put_contents($archivo, $datos, FILE_APPEND); // FILE_APPEND agrega los datos al final del archivo
+
+        $this->index();
+    }
+
+    public function procesarLogin(){
+        // Recoger los datos del formulario
+        $email = $_POST['inputEmail'];
+        $password = $_POST['inputPassword'];
+        $archivo = __DIR__ . "/../../login.txt";
+
+        if (!file_exists($archivo)) {
+            echo "⚠️ Archivo de usuarios no encontrado.";
+            return;
+        }
+    
+        $lineas = file($archivo, FILE_IGNORE_NEW_LINES | FILE_SKIP_EMPTY_LINES);
+        $credencialesValidas = false;
+    
+        foreach ($lineas as $linea) {
+            list($id, $emailArchivo, $passArchivo, $nombre, $apellido) = explode('|', trim($linea));
+            if ($email === $emailArchivo && $password === $passArchivo) {
+                $credencialesValidas = true;
+                break;
+            }
+        }
+    
+        if ($credencialesValidas) {
+            $this->index();
+            exit;
+        } else {
+            echo "❌ Email o contraseña incorrectos.";
+        }
+    }
+
+    public function procesarRegistro(){
+        // Recoger los datos del formulario
+        $nombre = $_POST['inputNombre'];
+        $apellido = $_POST['inputApellido'];
+        $email = $_POST['inputEmail'];
+        $password = $_POST['inputPassword'];
+        $confirmarPassword = $_POST['inputConfirmarPassword'];
+        if($password !== $confirmarPassword){
+            echo "⚠️ Las contraseñas no coinciden.";
+            return;
+        }
+
+        // Ruta del archivo de texto donde se guardarán los datos
+        $archivo = __DIR__ . "/../../login.txt";
+        
+        $id = 1; // Si es el primero
+        if (file_exists($archivo)) {
+            $lineas = file($archivo, FILE_IGNORE_NEW_LINES | FILE_SKIP_EMPTY_LINES);
+            $id = count($lineas) + 1;
+        }
+        // Formatear los datos en un texto legible
+        $datos = "$id|$email|$password|$nombre|$apellido\n";
         
         // Guardar los datos en el archivo de texto
         file_put_contents($archivo, $datos, FILE_APPEND); // FILE_APPEND agrega los datos al final del archivo
